Move Span definition to its own file
And prepare for more variants
