+++ dhall_syntax/src/core/mod.rs
@@ -4,6 +4,8 @@ mod import;
 pub use import::*;
 mod label;
 pub use label::*;
+mod span;
+pub use span::*;
 mod text;
 pub use text::*;
 pub mod context;
